dev-requirements.txt
-r requirements.txt
ipython
ipdb
django-dotenv
factory-boy>=2.5,<3.0